Restructure createDraggable so we can use if for popups as well
