+++ src/react/components/window/Flyout.jsx
@@ -2,6 +2,7 @@ import React, { useCallback, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { CloseCircleFilled } from '@ant-design/icons';
+import { createDraggable } from './util';
 
 const Container = styled.div`
     position: absolute;
@@ -49,62 +50,11 @@ const ToolsContainer = styled.div`
     }
 `;
 
-const getAvailableWidth = () => {
-    // width of <body>
-    return document.body.clientWidth ||
-        // width of <html>
-        document.documentElement.clientWidth ||
-        // window's width
-        window.innerWidth;
-};
-const getAvailableHeight = () => {
-    // height of <body>
-    return document.body.clientHeight ||
-        // height of <html>
-        document.documentElement.clientHeight ||
-        // window's height
-        window.innerHeight;
-};
 
 export const Flyout = ({title = '', children, onClose}) => {
     const [position, setPosition] = useState({ x: 210, y: 30 });
     const elementRef = useRef();
-    
-    const onMouseDown = useCallback(
-        (event) => {
-            const element = elementRef.current;
-            let width = 50;
-            let height = 30;
-            if (element) {
-                const bounds = element.getBoundingClientRect();
-                width = bounds.width;
-                height = bounds.height;
-            }
-            const availableWidth = getAvailableWidth();
-            const availableHeight = getAvailableHeight();
-            const onMouseMove = (event) => {
-                // prevents text selection from other elements while dragging
-                event.preventDefault();
-                position.x += event.movementX;
-                position.y += event.movementY;
-                const outOfScreen = (position.x < 0 || position.x + width > availableWidth) || (position.y < 0 || position.y + height > availableHeight);
-                const element = elementRef.current;
-                if (!outOfScreen && element) {
-                    // don't make the actual move if we would move off-screen or we don't get an element to move
-                    element.style.transform = `translate(${position.x}px, ${position.y}px)`;
-                }
-                // update state anyway so the flyout will start moving at/jump to where the mouse cursor re-enters the viewport
-                setPosition(position);
-            };
-            const onMouseUp = () => {
-                document.removeEventListener("mousemove", onMouseMove);
-                document.removeEventListener("mouseup", onMouseUp);
-            };
-            document.addEventListener("mousemove", onMouseMove);
-            document.addEventListener("mouseup", onMouseUp);
-        },
-        [position, setPosition, elementRef]
-    );
+    const onMouseDown = useCallback((event) => createDraggable(position, setPosition, elementRef), [position, setPosition, elementRef]);
     /*
     Other tools for toolcontainer:
         <div className="oskari-flyouttool-help"></div>
